+++ eznlp/__init__.py
@@ -2,9 +2,10 @@
 __version__ = '0.0.3'
 
 from .token import Token, TokenSequence, custom_spacy_tokenizer
-from .config import Config, ConfigList, ConfigDict
-from .config import TokenConfig, CharConfig, EnumConfig, ValConfig, EmbedderConfig
-from .config import EncoderConfig
-from .config import PreTrainedEmbedderConfig
+from .config import ConfigList, ConfigDict
+from .char_encoders import CharConfig
+from .embedders import TokenConfig, EnumConfig, ValConfig, EmbedderConfig
+from .encoders import EncoderConfig
+from .pretrained_embedders import PreTrainedEmbedderConfig
 from .optim_utils import count_trainable_params, build_param_groups_with_keyword2lr, check_param_groups_no_missing
 from .trainers import disp_running_info
