+++ eznlp/sequence_tagging/__init__.py
@@ -2,7 +2,7 @@
 from .transitions import ChunksTagsTranslator, SchemeTranslator
 from .raw_data import parse_conll_file
 from .datasets import SequenceTaggingDataset
-from .config import DecoderConfig, TaggerConfig
-from .taggers import Tagger
+from .decoders import DecoderConfig
+from .taggers import TaggerConfig
 from .trainers import SequenceTaggingTrainer
 from .metrics import precision_recall_f1_report
